Split SvgTextBase
